requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.7.1
sentence-transformers==3.0.0
qdrant-client==1.9.1
langchain-text-splitters==0.2.2
pymupdf==1.24.5
rank-bm25==0.2.2
aiofiles==23.2.1
